Update tests and build scripts. Add Travis
