+++ Library/src/androidTest/java/info/metadude/android/typedpreferences/BooleanPreferenceTest.java
@@ -8,28 +8,28 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
+import org.robolectric.annotation.Config;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
+@Config(emulateSdk = 18)
 @RunWith(RobolectricTestRunner.class)
-public class FloatPreferenceTest {
+public class BooleanPreferenceTest {
 
     protected static final String PREFERENCES_KEY =
-            "info.metadude.android.typedpreferences.TEST_KEY_FLOAT";
-    protected FloatPreference mPreference;
-    protected final float mTestValue = 23.23f;
-    protected final float mDefaultValue = 42.42f;
-    protected final float mDelta = 0f;
+            "info.metadude.android.typedpreferences.TEST_KEY_BOOLEAN";
+    protected BooleanPreference mPreference;
+    protected final boolean mTestValue = true;
+    protected final boolean mDefaultValue = false;
 
     @Before
     public void beforeEach() {
         final SharedPreferences sharedPreferences = PreferenceManager
                 .getDefaultSharedPreferences(Robolectric.application);
-        mPreference = new FloatPreference(sharedPreferences, PREFERENCES_KEY, mDefaultValue);
+        mPreference = new BooleanPreference(sharedPreferences, PREFERENCES_KEY, mDefaultValue);
     }
 
     @Test
@@ -41,8 +41,8 @@ public class FloatPreferenceTest {
     public void expect_Preference_To_EqualDefaultValue() throws Exception {
         final SharedPreferences sharedPreferences = PreferenceManager
                 .getDefaultSharedPreferences(Robolectric.application);
-        FloatPreference preference = new FloatPreference(sharedPreferences, PREFERENCES_KEY);
-        assertEquals(preference.get(), FloatPreference.DEFAULT_VALUE_VALUE, mDelta);
+        BooleanPreference preference = new BooleanPreference(sharedPreferences, PREFERENCES_KEY);
+        assertEquals(preference.get(), BooleanPreference.DEFAULT_VALUE_VALUE);
     }
 
     @Test
@@ -52,36 +52,33 @@ public class FloatPreferenceTest {
 
     @Test
     public void expect_Preference_To_BeSet() throws Exception {
-        mPreference.set(mTestValue);
+        mPreference.set(false);
         assertTrue(mPreference.isSet());
     }
 
     @Test
     public void expect_Preference_To_EqualDefaultValue_BeforeBeingSet() throws Exception {
-        final float value = mPreference.get();
-        assertEquals(value, mDefaultValue, mDelta);
-        assertNotEquals(value, mTestValue);
+        assertFalse(mPreference.get());
+        assertEquals(mPreference.get(), mDefaultValue);
     }
 
     @Test
     public void expect_Preference_To_EqualValue() throws Exception {
         mPreference.set(mTestValue);
-        final float value = mPreference.get();
-        assertEquals(value, mTestValue, mDelta);
+        assertTrue(mPreference.get());
     }
 
     @Test
     public void expect_Preference_NotTo_EqualValue() throws Exception {
-        mPreference.set(77.77f);
-        final float value = mPreference.get();
-        assertNotEquals(value, mTestValue, mDelta);
+        mPreference.set(false);
+        assertFalse(mPreference.get());
     }
 
     @Test
     public void expect_Preference_To_BeUnset() throws Exception {
         mPreference.set(mTestValue);
         mPreference.delete();
-        assertNotEquals(mPreference.get(), mTestValue);
+        assertFalse(mPreference.get());
     }
 
 }
