+++ Library/src/androidTest/java/info/metadude/android/typedpreferences/FloatPreferenceTest.java
@@ -8,6 +8,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
+import org.robolectric.annotation.Config;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -15,21 +16,22 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
+@Config(emulateSdk = 18)
 @RunWith(RobolectricTestRunner.class)
-public class IntPreferenceTest {
+public class FloatPreferenceTest {
 
     protected static final String PREFERENCES_KEY =
-            "info.metadude.android.typedpreferences.TEST_KEY_INT";
-    protected IntPreference mPreference;
-    protected final int mTestValue = 42;
-    protected final int mDefaultValue = 23;
-    protected final int mDelta = 0;
+            "info.metadude.android.typedpreferences.TEST_KEY_FLOAT";
+    protected FloatPreference mPreference;
+    protected final float mTestValue = 23.23f;
+    protected final float mDefaultValue = 42.42f;
+    protected final float mDelta = 0f;
 
     @Before
     public void beforeEach() {
         final SharedPreferences sharedPreferences = PreferenceManager
                 .getDefaultSharedPreferences(Robolectric.application);
-        mPreference = new IntPreference(sharedPreferences, PREFERENCES_KEY, mDefaultValue);
+        mPreference = new FloatPreference(sharedPreferences, PREFERENCES_KEY, mDefaultValue);
     }
 
     @Test
@@ -41,9 +43,8 @@ public class IntPreferenceTest {
     public void expect_Preference_To_EqualDefaultValue() throws Exception {
         final SharedPreferences sharedPreferences = PreferenceManager
                 .getDefaultSharedPreferences(Robolectric.application);
-        IntPreference preference = new IntPreference(sharedPreferences, PREFERENCES_KEY);
-        final int value = preference.get();
-        assertEquals(value, IntPreference.DEFAULT_VALUE_VALUE);
+        FloatPreference preference = new FloatPreference(sharedPreferences, PREFERENCES_KEY);
+        assertEquals(preference.get(), FloatPreference.DEFAULT_VALUE_VALUE, mDelta);
     }
 
     @Test
@@ -59,29 +60,30 @@ public class IntPreferenceTest {
 
     @Test
     public void expect_Preference_To_EqualDefaultValue_BeforeBeingSet() throws Exception {
-        final int value = mPreference.get();
-        assertEquals(value, mDefaultValue);
+        final float value = mPreference.get();
+        assertEquals(value, mDefaultValue, mDelta);
         assertNotEquals(value, mTestValue);
     }
 
     @Test
     public void expect_Preference_To_EqualValue() throws Exception {
         mPreference.set(mTestValue);
-        final int value = mPreference.get();
-        assertEquals(value, mTestValue);
+        final float value = mPreference.get();
+        assertEquals(value, mTestValue, mDelta);
     }
 
     @Test
     public void expect_Preference_NotTo_EqualValue() throws Exception {
-        mPreference.set(55);
-        assertNotEquals(mPreference.get(), mTestValue, mDelta);
+        mPreference.set(77.77f);
+        final float value = mPreference.get();
+        assertNotEquals(value, mTestValue, mDelta);
     }
 
     @Test
     public void expect_Preference_To_BeUnset() throws Exception {
         mPreference.set(mTestValue);
         mPreference.delete();
-        assertNotEquals(mPreference.get(), mTestValue, mDelta);
+        assertNotEquals(mPreference.get(), mTestValue);
     }
 
 }
