+++ TypedPreferencesDemo/build.gradle
@@ -1,12 +1,12 @@
-apply plugin: 'android'
+apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 19
-    buildToolsVersion "19.1.0"
+    compileSdkVersion 20
+    buildToolsVersion "20.0.0"
 
     defaultConfig {
         minSdkVersion 8
-        targetSdkVersion 19
+        targetSdkVersion 20
         versionCode 1
         versionName "1.1.0"
     }
@@ -19,7 +19,6 @@ android {
 }
 
 dependencies {
-    compile 'info.metadude.android:typed-preferences:1.1.+'
-    compile 'com.android.support:appcompat-v7:+'
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile project(':Library')
+    compile 'com.android.support:appcompat-v7:20.0.0'
 }
