+++ build.gradle
@@ -3,24 +3,15 @@
 buildscript {
     repositories {
         mavenCentral()
-        mavenLocal()
-        maven {
-            url 'https://oss.sonatype.org/content/repositories/snapshots'
-        }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:0.11.+'
-        classpath 'com.github.dcendents:android-maven-plugin:1.0'
-        classpath 'com.novoda.gradle:robolectric-plugin:0.0.1'
+        classpath 'com.android.tools.build:gradle:0.12.2'
     }
 }
 
 allprojects {
     repositories {
         mavenCentral()
-        maven {
-            url 'https://oss.sonatype.org/content/repositories/snapshots'
-        }
     }
 }
