+++ gradle.properties
@@ -1,0 +1,4 @@
+# Dummy properties so that anyone can compile the library
+sonatypeRepo=""
+sonatypeUserName=""
+sonatypePassword=""
